MVCContactBook/ContactBookApi/Controllers/StateController.cs
using ContactBookApi.Services.Contract;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;

namespace ContactBookApi.Controllers
{
    [Route("api/[controller]")]
    [ApiController]
    public class StateController : ControllerBase
    {
        private readonly IStateService _stateService;

        public StateController(IStateService stateService)
        {
            _stateService = stateService;
        }

        [HttpGet("GetStateByCountry/{id}")]
        public IActionResult GetStatesByCountryId(int id)
        {
            var response = _stateService.GetStatesByCountryId(id);

            if (response.Success)
            {
                return Ok(response);
            }
            else
            {
                return NotFound(response);
            }
        }

        [HttpGet("GetAllStates")]

        public IActionResult GetAllStates()
        {
            var response = _stateService.GetStates();
            if (!response.Success)
            {
                return BadRequest(response);
            }
            return Ok(response);
        }
    }
}
